# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel Planner API documentation",
        "description": "API Documentation for Travel Planner",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/accommodation": {
            "get": {
                "tags": [
                    "Accommodation"
                ],
                "summary": "Get all accommodations for logged in user",
                "description": "Get all accommodations for logged in user",
                "operationId": "get_accommodation",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Accommodation"
                ],
                "summary": "Create a accommodation for a destination",
                "description": "Create accommodation",
                "operationId": "create_accommodation",
                "parameters": [
                    {
                        "name": "destination_id",
                        "in": "query",
                        "description": "required | exists:destinations,id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "accommodation_name",
                        "in": "query",
                        "description": "required_with:accommodations | string | max:100",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "checkin_time",
                        "in": "query",
                        "description": "date_format:Y-m-d H:i | after:today",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "checkout_time",
                        "in": "query",
                        "description": "date_format:Y-m-d H:i | after:start_date",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "cost",
                        "in": "query",
                        "description": "numeric | min:0",
                        "schema": {
                            "type": "decimal"
                        }
                    },
                    {
                        "name": "booking_id",
                        "in": "query",
                        "description": "string | max:20",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/accommodation/update": {
            "post": {
                "tags": [
                    "Accommodation"
                ],
                "summary": "Update accommodation",
                "description": "Update accommodation",
                "operationId": "update_accommodation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "required | exists:accommodations,id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "accommodation_name",
                        "in": "query",
                        "description": "required_with:accommodations | string | max:100",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "checkin_time",
                        "in": "query",
                        "description": "date_format:Y-m-d H:i | after:today",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "checkout_time",
                        "in": "query",
                        "description": "date_format:Y-m-d H:i | after:start_date",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "cost",
                        "in": "query",
                        "description": "numeric | min:0",
                        "schema": {
                            "type": "decimal"
                        }
                    },
                    {
                        "name": "booking_id",
                        "in": "query",
                        "description": "string | max:20",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/accommodation/delete": {
            "post": {
                "tags": [
                    "Accommodation"
                ],
                "summary": "Delete accommodation",
                "description": "Delete accommodation",
                "operationId": "delete_accommodation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "required | exists:accommodations,id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/batch/accommodation": {
            "post": {
                "tags": [
                    "Accommodation"
                ],
                "summary": "Create multiple accommodation for a destination",
                "description": "Create accommodation",
                "operationId": "batch_create_accommodation",
                "parameters": [
                    {
                        "name": "destination_id",
                        "in": "query",
                        "description": "required | exists:destinations,id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "body",
                        "in": "query",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "properties": {
                                "accommodation": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "accommodation_name": {
                                                "type": "string"
                                            },
                                            "checkin_time": {
                                                "type": "string"
                                            },
                                            "checkout_time": {
                                                "type": "string"
                                            },
                                            "cost": {
                                                "type": "integer"
                                            },
                                            "booking_id": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/accommodation/search": {
            "get": {
                "tags": [
                    "Accommodation"
                ],
                "summary": "Find accommodation by city",
                "description": "Find accommodation by city",
                "operationId": "search_accommodation",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "description": "required | string",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}